rename KiiPromise > PromiseWrapper
